requirements.txt
# Gabarita-AI - Dependências
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Login==0.6.3
requests==2.31.0
openai==1.3.0
firebase-admin==6.2.0
mercadopago==2.2.1
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
urllib3==2.0.7
PyJWT==2.8.0
cryptography==41.0.7
grpcio==1.59.0
grpcio-status==1.59.0
protobuf==4.24.4
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
pyparsing==3.1.1
rsa==4.9
six==1.16.0
cachetools==5.3.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
